+++ packages/web3-eth-contract/src/contract.ts
@@ -42,6 +42,8 @@ import {
 	TransactionMiddleware,
 } from 'web3-eth';
 import {
+	decodeMethodReturn,
+	decodeMethodParams,
 	encodeEventSignature,
 	encodeFunctionSignature,
 	decodeContractErrorData,
@@ -99,12 +101,7 @@ import {
 	ValidationSchemaInput,
 	Web3ValidatorError,
 } from 'web3-validator';
-import {
-	decodeMethodReturn,
-	decodeMethodParams,
-	encodeEventABI,
-	encodeMethodABI,
-} from './encoding.js';
+import { encodeEventABI, encodeMethodABI } from './encoding.js';
 import { ContractLogsSubscription } from './contract_log_subscription.js';
 import {
 	ContractEventOptions,
@@ -1026,7 +1023,7 @@ export class Contract<Abi extends ContractAbi>
 				`The ABI for the provided method signature ${methodSignature} was not found.`,
 			);
 		}
-		return { ...decodeMethodParams(abi, data), __method__: jsonInterfaceMethodToString(abi) };
+		return decodeMethodParams(abi, data);
 	}
 
 	private _parseAndSetJsonInterface(
